Scatterplot overhaul progress.  Should just need to fix bootstrap tooltip loading and it's good to go.
