+++ client/package.json
@@ -47,6 +47,7 @@
     "build-toolshed": "grunt --app=toolshed",
     "jshint": "jshint --exclude='galaxy/scripts/libs/**' galaxy/scripts/**/*.js",
     "build-charts": "webpack -p --config ../config/plugins/visualizations/charts/webpack.config.js",
+    "build-scatterplot": "NODE_PATH=./node_modules webpack -d --config ../config/plugins/visualizations/scatterplot/webpack.config.js",
     "prettier": "prettier --write --tab-width 4 --print-width 120 \"galaxy/scripts/{,!(libs)/**/}/*.js\""
   },
   "devDependencies": {
