+++ config/plugins/visualizations/scatterplot/static/worker-stats.js
@@ -1,5 +1,5 @@
-onmessage = function( event ){
-    importScripts( 'numeric-column-stats.js' );
-    postMessage( numericColumnStats( event.data.data, event.data.keys ) );
+onmessage = function(event) {
+    importScripts("numeric-column-stats.js");
+    postMessage(numericColumnStats(event.data.data, event.data.keys));
     self.close();
 };
